optee_client: [OL 9.4] [New OS] optee-client compilation error
Changes to allow the optee_client to build natively
for Oracle Linux 9.4. We want to build natively on all
our OP-TEE Client platforms.

These changes work for Debian/Ubuntu as well.

Changes made:

1) Removed the CROSS_COMPILE flag as the OP-TEE Client is always built natively.
   Also removed associated unnecessary Debian patch for this.
(2) Updated Cmake flags so that the created artifacts (for OCI) are not

    placed in the redhat-linux-build directory, as that causes cmake install
    to fail installing the built artifacts.
(3) Updated the tag to 4.2.3 for these changes.

RM #4196944